Add setFrequency fn, and pitch slider to demo.
Now we have two sliders in the demo! One controls volume, the other
pitch/frequency. This generally seems to work well, though I've observed
a bug in the way volume works while muted. If I change the volume slider
while the noisemaker is muted, it starts playing at the new volume. The
mute checkbox stays checked, and the isMuted value is still "true."
Fixing that next.